Add scheduling and DaemonSet workshop
Added a new workshop on Scheduling and DaemonSets (workshop-08) to the course material. Created several YAML configuration files for the workshop-08 directory, detailing the creation of specific pods and a DaemonSet. The `.gitignore` file was updated to add 'ingress' and links to the new workshop were provided in the main README file.